backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.4.0
google-cloud-secret-manager==2.18.1
google-cloud-storage==2.13.0
google-cloud-logging==3.8.0
pydantic-settings==2.1.0
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1
requests==2.31.0
aiofiles==23.2.1
python-dotenv==1.0.0
prometheus-client==0.19.0